Separate syntax receiver to new file
