+++ src/libraries/System.Text.Json/System.Text.Json.SourceGeneration/System.Text.Json.SourceGeneration.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
@@ -31,6 +31,7 @@
     <Compile Include="ReflectionUtils\RoslynExtensions.cs" />
     <Compile Include="ReflectionUtils\TypeWrapper.cs" />
     <Compile Include="JsonSourceGenerator.cs" />
+    <Compile Include="JsonSerializableSyntaxReceiver.cs" />
   </ItemGroup>
 
 </Project>
